Fix discrepancy in history file variable name
The string history file had 2 different variable names.  In default values and the main code, it was named "sm_history".  In the command line parsing, it was named "string_method_history".  In the previous version, command line parsing did not change the history file name and therefore was always using the default name.  In the updated code, all variables for the history file are changed to "string_method_history".